nlp/requirements.txt
pymongo
neo4j
python-dotenv
spacy
scispacy
